@@ -10,15 +10,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.2.0",
+    "@angular/animations": "~14.3.0",
     "@angular/cdk": "^13.3.2",
-    "@angular/common": "~13.2.0",
-    "@angular/compiler": "~13.2.0",
-    "@angular/core": "~13.2.0",
-    "@angular/forms": "~13.2.0",
-    "@angular/platform-browser": "~13.2.0",
-    "@angular/platform-browser-dynamic": "~13.2.0",
-    "@angular/router": "~13.2.0",
+    "@angular/common": "~14.3.0",
+    "@angular/compiler": "~14.3.0",
+    "@angular/core": "~14.3.0",
+    "@angular/forms": "~14.3.0",
+    "@angular/platform-browser": "~14.3.0",
+    "@angular/platform-browser-dynamic": "~14.3.0",
+    "@angular/router": "~14.3.0",
     "@microsoft/signalr": "^6.0.8",
     "@ngx-translate/core": "^14.0.0",
     "@ngx-translate/http-loader": "^7.0.0",
@@ -40,7 +40,7 @@
   "devDependencies": {
     "@angular-devkit/build-angular": "^18.1.1",
     "@angular/cli": "^18.1.1",
-    "@angular/compiler-cli": "~13.2.0",
+    "@angular/compiler-cli": "~14.3.0",
     "@types/jasmine": "~3.10.0",
     "@types/node": "^12.11.1",
     "autoprefixer": "^10.4.4",
@@ -53,6 +53,6 @@
     "karma-jasmine-html-reporter": "~1.7.0",
     "postcss": "^8.4.12",
     "tailwindcss": "^3.0.23",
-    "typescript": "~4.5.2"
+    "typescript": "~4.8.4"
   }
 }
